test: Split sender tests out to a separate module under unittests
